Abaqus reduce pre

+++ coupling_components/solver_wrappers/abaqus/v614.py
@@ -31,6 +31,9 @@ class SolverWrapperAbaqus614(Component):
         self.dir_csm = join(os.getcwd(), self.settings['working_directory'])  # *** alternative for getcwd?
         self.env = tools.get_solver_env(__name__, self.dir_csm)
         self.check_software()
+        self.dir_vault = Path(self.dir_csm) / 'vault'
+        self.dir_vault.mkdir(exist_ok=True)
+        self.vault_suffixes = ['023', 'com', 'dat', 'mdl', 'msg', 'odb', 'prt', 'res', 'sim', 'stt']
         path_src = os.path.realpath(os.path.dirname(__file__))
         self.logfile = 'abaqus.log'
 
@@ -359,9 +362,31 @@ class SolverWrapperAbaqus614(Component):
                     f'cpus={self.cores} output_precision=full interactive >> {self.logfile} 2>&1'
                 subprocess.run(cmd, shell=True, cwd=self.dir_csm, executable='/bin/bash', env=self.env)
             else:
-                cmd = f'abaqus job=CSM_Time{self.timestep} oldjob=CSM_Time{self.timestep - 1} input=CSM_Restart ' \
-                    f'cpus={self.cores} output_precision=full interactive >> {self.logfile} 2>&1'
-                subprocess.run(cmd, shell=True, cwd=self.dir_csm, executable='/bin/bash', env=self.env)
+                if self.iteration == 1:
+                    # run datacheck and store generated files safely
+                    cmd = f'abaqus datacheck job=CSM_Time{self.timestep} oldjob=CSM_Time{self.timestep - 1} ' \
+                        f'input=CSM_Restart cpus={self.cores} output_precision=full interactive ' \
+                        f'>> {self.logfile} 2>&1'
+                    subprocess.run(cmd, shell=True, cwd=self.dir_csm, executable='/bin/bash', env=self.env)
+                    for suffix in self.vault_suffixes:
+                        from_file = Path(self.dir_csm) / f'CSM_Time{self.timestep}.{suffix}'
+                        to_file = self.dir_vault / f'CSM_Time{self.timestep}.{suffix}'
+                        shutil.copy(from_file, to_file)
+                    # run continue
+                    cmd = f'abaqus continue job=CSM_Time{self.timestep} oldjob=CSM_Time{self.timestep - 1} ' \
+                        f'input=CSM_Restart cpus={self.cores} output_precision=full interactive ' \
+                        f'>> {self.logfile} 2>&1'
+                    subprocess.run(cmd, shell=True, cwd=self.dir_csm, executable='/bin/bash', env=self.env)
+                else:
+                    # run continue using previously stored simulation files
+                    for suffix in self.vault_suffixes:
+                        from_file = self.dir_vault / f'CSM_Time{self.timestep}.{suffix}'
+                        to_file = Path(self.dir_csm) / f'CSM_Time{self.timestep}.{suffix}'
+                        shutil.copy(from_file, to_file)
+                    cmd = f'abaqus continue job=CSM_Time{self.timestep} oldjob=CSM_Time{self.timestep - 1} ' \
+                        f'input=CSM_Restart cpus={self.cores} output_precision=full interactive ' \
+                        f'>> {self.logfile} 2>&1'
+                    subprocess.run(cmd, shell=True, cwd=self.dir_csm, executable='/bin/bash', env=self.env)
 
             # check log for completion and or errors
             subprocess.run(f'tail -n 10 {self.logfile} > Temp_log.coco', shell=True, cwd=self.dir_csm,
@@ -422,16 +447,20 @@ class SolverWrapperAbaqus614(Component):
 
     def finalize_solution_step(self):
         super().finalize_solution_step()
-        if self.save_interval == 0 or (self.timestep - 1) % self.save_interval != 0:
-            to_be_removed_suffix = ['.com', '.dat', '.mdl', '.msg', '.odb', '.prt', '.res', '.sim', '.sta', '.stt',
-                                    'Surface*Cpu0Input.dat', 'Surface*Output.dat']
-            cmd = ''
-            for suffix in to_be_removed_suffix:
-                cmd += f'rm CSM_Time{self.timestep - 1}{suffix}; '
-            subprocess.run(cmd, shell=True, cwd=self.dir_csm, executable='/bin/bash', env=self.env)
+        if self.timestep > 0:
+            if self.save_interval == 0 or (self.timestep - 1) % self.save_interval != 0:
+                to_be_removed_suffix = ['.com', '.dat', '.mdl', '.msg', '.odb', '.prt', '.res', '.sim', '.sta', '.stt',
+                                        'Surface*Cpu0Input.dat', 'Surface*Output.dat']
+                cmd = ''
+                for suffix in to_be_removed_suffix:
+                    cmd += f'rm CSM_Time{self.timestep - 1}{suffix}; '
+                subprocess.run(cmd, shell=True, cwd=self.dir_csm, executable='/bin/bash', env=self.env)
+            for f in self.dir_vault.iterdir():
+                f.unlink()  # empty vault
 
     def finalize(self):
         super().finalize()
+        self.dir_vault.rmdir()
 
     def get_interface_input(self):
         return self.interface_input
